% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{animal_info}
\alias{animal_info}
\title{Get information about an animal}
\usage{
learnfun::animal_info(name,, where, age)
}
\arguments{
\item{name}{\link{character} The animal's name}

\item{where}{\link{chracter} The location where the animal lives}

\item{lifespan}{\link{numeric} The lifespan of the animal}
}
\value{
Returns a \link{tibble} containing animal information
}
\description{
\code{animal_info()} prints out information about an animal
}
