% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{number_and_word}
\alias{number_and_word}
\title{Say a word and a number}
\usage{
learnfun::number_and_word(number, word)
}
\arguments{
\item{number}{The number you want to print}

\item{word}{The word you want to print}
}
\value{

}
\description{
\code{number_and_word()} prints out a number and a word
}
