% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rankitems}
\alias{rankitems}
\title{Rank food items}
\usage{
learnfun::rankitems(fav, second, hated)
}
\arguments{
\item{fav}{Your favourite food}

\item{second}{Your second favourite food}

\item{hated}{A food you hate}
}
\value{

}
\description{
\code{rankitems()} prints out a list of your favourite food, you second favourite
food, and a food you hate.
}
